Modularised pipeline
Former-commit-id: 6e9695a5a4e085cce7f99ade368a3ade93fbc4ba
