+++ module_1/06-genomic_diversity/genomic_het_table.tsv
@@ -1,0 +1,5 @@
+pop1	numerator	denominator	het
+FG04258	532.0	3187256.0	0.00016691473794386142
+MP_AC48	240.0	8899650.0	0.00002696735264869967
+MP_BR24	482.0	2192266.0	0.00021986383039284467
+MP_D48	740.0	4956370.0	0.00014930281637569431
