+++ module_1/Snakefile
@@ -23,8 +23,9 @@ include: "rules/03-vcftools_filter.smk"
 include: "rules/04-bcftools_sort.smk"
 include: "rules/05-bcftools_merge.smk"
 include: "rules/06-python_filter.smk"
+include: "rules/07-piawka_het.smk"
 
 #######################################################################################
 rule all:
 	input:
-		config['python_filter']['output_dir']  + 'all_merged_filtered.vcf.gz'
+		config['piawka_agg']['output_dir']  + 'genomic_het_table.tsv'
